This commit is a huge change in how the plugin works, for now, it will work inside craftbukkit's jar, why? well, tons of reasons, first of all, when I hook the player with my custom playerConnection class, everything starts to behave weirdly, it mess with player movement, teleport, everything, and cut the compatibility with tons of usefull plugins, probably I'm doing something wrong, but I can't find the reason.
Second and most important reason, the goal of this plugin is and always will be the performance, using the old method, the obfuscation process takes around 5 MS to 10 MS (20 MS to 30MS with engine mode 2) to be completed and the chunk be sent to the player, when I do the obfuscation inside the craftbukkit jar it is made in less then 1 MS with engine mode 0 or 1 and 3 to 5MS with engine mode 2. What I mean is, when I do the obfuscation inside craftbukkit's jar, it will take much less resources and it will be completed way much faster, avoiding tps drops and slow chunk sending.

The cache was also removed in this commit, the current cache system generates too many files and is not close to usefull on server with a significant amount of players because of the Too many open files error, in way to keep the cache lightweight, I decided to use a lot of small files, which was not a very good ideia, and as I said above, the obfuscation process is way too fast, the cache will just slow things down.

Well, a rollback commit can always be expected, I'm testing it yet, but until now, it is a way better solution.

But what about bukkit updates? Not a problem at all, it only modifies one single line of the Packet51MapChunk and this file is not updated very often.